Block 9 MasterWorks
Block 9 MasterWorks is a collection of 100 of the some of the greatest Art Masterpieces ever painted, all immutably inscribed on Block 9 Sats for Digital Ownership. Owning a Block 9 MasterWorks Digital Asset also serves as a MasterPass for Exclusive Access on every collection we ever drop. It also carries with it perks, such as substantial discounts on our Ordinals, Merchandise, GenesisVIP holders channel in discord, Airdrops, and more...